requirements.txt
Pillow>=8.0.0
psd-tools>=1.9.0
